requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
PyYAML
xgboost
dagshub
mlflow
fastapi
flask
##-e .